+++ KtorApi-WhiskerWatch/src/main/kotlin/com/example/routing/MapsRoutes.kt
@@ -1,12 +1,14 @@
 package com.example.routing
 
 import com.example.dto.MapsCreateDto
+import com.example.models.users.Rol
 import com.example.services.maps.MapsService
 import com.github.michaelbull.result.onFailure
 import com.github.michaelbull.result.onSuccess
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
+import io.ktor.server.auth.jwt.*
 import io.ktor.server.plugins.requestvalidation.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
@@ -57,9 +59,15 @@ fun Application.mapRoutes(){
                     try {
                         val id = call.parameters["id"].toString()
                         val put = call.receive<MapsCreateDto>()
-                        service.updateMap(put, id)
-                            .onSuccess { call.respond(HttpStatusCode.Created, it) }
-                            .onFailure { call.respond(HttpStatusCode.NotFound, it.message) }
+                        val auth = call.principal<JWTPrincipal>()
+                        val rol = Rol.valueOf(auth?.payload?.getClaim("rol").toString().replace("\"", ""))
+                        if(rol != Rol.ADMIN){
+                            call.respond(HttpStatusCode.Unauthorized, "No tiene permisos para realizar esta acción")
+                        }else {
+                            service.updateMap(put, id)
+                                .onSuccess { call.respond(HttpStatusCode.Created, it) }
+                                .onFailure { call.respond(HttpStatusCode.NotFound, it.message) }
+                        }
                     }catch (e: RequestValidationException){
                         call.respond(HttpStatusCode.BadRequest, e.reasons.toString())
                     }
@@ -68,17 +76,29 @@ fun Application.mapRoutes(){
                 delete("/{id}"){
                     logger.info { "Delete Map" }
                     val id = call.parameters["id"].toString()
-                    service.deleteMap(id)
-                        .onSuccess { call.respond(HttpStatusCode.NoContent) }
-                        .onFailure { call.respond(HttpStatusCode.NotFound, it.message) }
+                    val auth = call.principal<JWTPrincipal>()
+                    val rol = Rol.valueOf(auth?.payload?.getClaim("rol").toString().replace("\"", ""))
+                    if( rol != Rol.ADMIN){
+                        call.respond(HttpStatusCode.Unauthorized, "No tiene permisos para realizar esta acción")
+                    }else {
+                        service.deleteMap(id)
+                            .onSuccess { call.respond(HttpStatusCode.NoContent) }
+                            .onFailure { call.respond(HttpStatusCode.NotFound, it.message) }
+                    }
                 }
 
                 delete("/adoption/{id}"){
                     logger.info { "Delete Map Association Adoption" }
                     val id = call.parameters["id"].toString()
-                    service.deleteMapsAdoption(id)
-                        .onSuccess { call.respond(HttpStatusCode.NoContent) }
-                        .onFailure { call.respond(HttpStatusCode.NotFound, it.message) }
+                    val auth = call.principal<JWTPrincipal>()
+                    val rol = Rol.valueOf(auth?.payload?.getClaim("rol").toString().replace("\"", ""))
+                    if(rol == Rol.USER){
+                        call.respond(HttpStatusCode.Unauthorized, "No tienes permisos para realizar esta acción")
+                    }else {
+                        service.deleteMapsAdoption(id)
+                            .onSuccess { call.respond(HttpStatusCode.NoContent) }
+                            .onFailure { call.respond(HttpStatusCode.NotFound, it.message) }
+                    }
                 }
 
             }
